# specs/001-quickflow-productivity/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QuickFlow API
  version: 1.0.0
  description: >
    REST API of QuickFlow (single-user personal productivity). Source of truth for the backend springdoc swagger
    and the generated typescript-angular client. Errors use RFC 9457 application/problem+json.
servers:
- {url: 'http://localhost:8080'}
...
paths:
...
  /api/learning-cards:
    get:
      tags: [Learning]
      operationId: listLearningCards
      responses:
        '200':
          description: Cards with milestones and notes, newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/LearningCard'}
    post:
      tags: [Learning]
      operationId: createLearningCard
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningCardRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/learning-cards/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    get:
      tags: [Learning]
      operationId: getLearningCard
      responses:
        '200':
          description: The card
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Learning]
      operationId: updateLearningCard
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningCardRequest'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Learning]
      operationId: deleteLearningCard
      responses:
        '204': {description: Deleted with milestones and notes}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/learning-cards/{id}/milestones:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Learning]
      operationId: addMilestone
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MilestoneRequest'}
      responses:
        '201':
          description: Added; returns the updated card
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/learning-cards/{id}/milestones/{milestoneId}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    - name: milestoneId
      in: path
      required: true
      schema: {type: integer, format: int64}
    put:
      tags: [Learning]
      operationId: updateMilestone
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MilestoneRequest'}
      responses:
        '200':
          description: Updated; returns the updated card
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Learning]
      operationId: deleteMilestone
      responses:
        '200':
          description: Removed; returns the updated card
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/learning-cards/{id}/notes:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    post:
      tags: [Learning]
      operationId: addNote
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NoteRequest'}
      responses:
        '201':
          description: Added; returns the updated card
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/learning-cards/{id}/notes/{noteId}:
    parameters:
    - name: id
      in: path
      required: true
      schema: {type: integer, format: int64}
    - name: noteId
      in: path
      required: true
      schema: {type: integer, format: int64}
    delete:
      tags: [Learning]
      operationId: deleteNote
      responses:
        '200':
          description: Removed; returns the updated card
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningCard'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        errors:
          type: array
          items: {$ref: '#/components/schemas/FieldError'}
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: {type: string}
        message: {type: string}
...
    LearningStatus:
      type: string
      enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
    LearningCardRequest:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        description:
          type: [string, 'null']
          maxLength: 2000
        status: {$ref: '#/components/schemas/LearningStatus'}
    MilestoneRequest:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 200}
        done: {type: boolean}
        targetDate:
          type: [string, 'null']
          format: date
    Milestone:
      type: object
      required: [id, title, done]
      properties:
        id: {type: integer, format: int64}
        title: {type: string}
        done: {type: boolean}
        targetDate:
          type: [string, 'null']
          format: date
        completedAt:
          type: [string, 'null']
          format: date-time
    NoteRequest:
      type: object
      required: [text]
      properties:
        text: {type: string, minLength: 1, maxLength: 5000}
    Note:
      type: object
      required: [id, text, createdAt]
      properties:
        id: {type: integer, format: int64}
        text: {type: string}
        createdAt: {type: string, format: date-time}
    LearningCard:
      type: object
      required: [id, title, status, createdAt, milestones, notes, milestonesTotal, milestonesDone, progressPercent]
      properties:
        id: {type: integer, format: int64}
        title: {type: string}
        description:
          type: [string, 'null']
        status: {$ref: '#/components/schemas/LearningStatus'}
        createdAt: {type: string, format: date-time}
        milestones:
          type: array
          items: {$ref: '#/components/schemas/Milestone'}
        notes:
          type: array
          items: {$ref: '#/components/schemas/Note'}
        milestonesTotal: {type: integer}
        milestonesDone: {type: integer}
        progressPercent: {type: integer, minimum: 0, maximum: 100}
...